renaming the productController file
