+++ .slingshot/agentHooks.json
@@ -1,0 +1,24 @@
+[
+  {
+    "id": "hook-1772429638832-288",
+    "name": "Code review",
+    "description": "Review the code for all coding standard",
+    "triggerType": "file_saved",
+    "filePaths": "src/*.js",
+    "prompt": "Analyze the recently saved file and provide a comprehensive code review with actionable feedback. Focus on:\n\n**Code Quality & Standards:**\n- Adherence to language-specific coding conventions and style guidelines\n- Proper naming conventions for variables, functions, and classes\n- Code structure, readability, and maintainability\n- Consistent formatting and indentation\n\n**Best Practices:**\n- Error handling and edge case management\n- Performance considerations and potential optimizations\n- Security vulnerabilities or concerns\n- Code reusability and DRY principles\n- Proper commenting and documentation\n\n**Technical Review:**\n- Logic correctness and potential bugs\n- Resource management (memory, file handles, connections)\n- Appropriate use of design patterns\n- Type safety and null checks where applicable\n\n**Output Format:**\nProvide specific, line-referenced feedback when possible. Categorize issues by severity (Critical, Major, Minor, Suggestion). Include positive observations alongside improvement recommendations. Keep comments constructive and educational, explaining the \"why\" behind each suggestion.\n\nIf no issues are found, acknowledge good practices observed in the code.",
+    "agentId": "coding-agent",
+    "parameters": {
+      "task": "Analyze the recently saved file and provide a comprehensive code review with actionable feedback. Focus on:\n\n**Code Quality & Standards:**\n- Adherence to language-specific coding conventions and style guidelines\n- Proper naming conventions for variables, functions, and classes\n- Code structure, readability, and maintainability\n- Consistent formatting and indentation\n\n**Best Practices:**\n- Error handling and edge case management\n- Performance considerations and potential optimizations\n- Security vulnerabilities or concerns\n- Code reusability and DRY principles\n- Proper commenting and documentation\n\n**Technical Review:**\n- Logic correctness and potential bugs\n- Resource management (memory, file handles, connections)\n- Appropriate use of design patterns\n- Type safety and null checks where applicable\n\n**Output Format:**\nProvide specific, line-referenced feedback when possible. Categorize issues by severity (Critical, Major, Minor, Suggestion). Include positive observations alongside improvement recommendations. Keep comments constructive and educational, explaining the \"why\" behind each suggestion.\n\nIf no issues are found, acknowledge good practices observed in the code.",
+      "files": [
+        "src/*.js"
+      ],
+      "context": "Review the code for all coding standard"
+    },
+    "enabled": false,
+    "createdBy": "user",
+    "tags": [],
+    "debounceDuration": "120000",
+    "createdAt": "2026-03-02T05:33:58.836Z",
+    "updatedAt": "2026-03-02T05:33:58.836Z"
+  }
+]
