+++ .slingshot/code-tagging/24d33d36-lock.json
@@ -1,0 +1,26 @@
+{
+  "backend/config/firebaseConfig.js": {
+    "filename": "firebaseConfig.js",
+    "path": "backend/config/firebaseConfig.js",
+    "version": "1.2",
+    "contextFile": null,
+    "chunks": {
+      "040c70": {
+        "user": "Karan Dutt",
+        "model": "claude-sonnet-4@20250514",
+        "timestamp": 1772795358,
+        "platform": "VisualStudioCode",
+        "product": "AIPP",
+        "productVersion": "2.2.0",
+        "source": "chat",
+        "type": "pasteCode",
+        "extraContext": {},
+        "hash": "1e81b1d8",
+        "messageId": "system_dc589330-194b-11f1-a9a0-6f3ba034df13",
+        "codeBlockId": "system_dc589330-194b-11f1-a9a0-6f3ba034df13-Codeblock-2",
+        "isActive": false
+      }
+    },
+    "lines": {}
+  }
+}
