+++ .slingshot/code-tagging/6c6e6c16-lock.json
@@ -19,8 +19,45 @@
         "messageId": "system_e2a3c6e0-124a-11f1-8c7f-2b7fea10d3af",
         "codeBlockId": "system_e2a3c6e0-124a-11f1-8c7f-2b7fea10d3af-Codeblock-1",
         "isActive": false
+      },
+      "5403bb": {
+        "user": "Karan Dutt",
+        "model": "claude-sonnet-4@20250514",
+        "timestamp": 1772457221,
+        "platform": "VisualStudioCode",
+        "product": "AIPP",
+        "productVersion": "2.2.0",
+        "source": "chat",
+        "type": "pasteCode",
+        "extraContext": {},
+        "hash": "c4950362",
+        "messageId": "system_7f05ceb0-1639-11f1-91f8-e9a0acd396d5",
+        "codeBlockId": "system_7f05ceb0-1639-11f1-91f8-e9a0acd396d5-Codeblock-2",
+        "isActive": true
+      },
+      "69a800": {
+        "user": "Karan Dutt",
+        "model": "claude-sonnet-4@20250514",
+        "timestamp": 1772696857,
+        "platform": "VisualStudioCode",
+        "product": "AIPP",
+        "productVersion": "2.2.0",
+        "source": "chat",
+        "type": "pasteCode",
+        "extraContext": {},
+        "hash": "cc105a04",
+        "messageId": "system_0ebfb960-1867-11f1-956e-13fc8e1aa9e0",
+        "codeBlockId": "system_0ebfb960-1867-11f1-956e-13fc8e1aa9e0-Codeblock-2",
+        "isActive": false
       }
     },
-    "lines": {}
+    "lines": {
+      "106": {
+        "chunkId": "5403bb",
+        "hash": "198da7d0",
+        "encoded": "const%20productId%20=%20req.params.id;",
+        "prevChunkIds": []
+      }
+    }
   }
 }
