+++ .slingshot/code-tagging/90925308-lock.json
@@ -1,0 +1,26 @@
+{
+  "backend/validators/product/productValidator.js": {
+    "filename": "productValidator.js",
+    "path": "backend/validators/product/productValidator.js",
+    "version": "1.2",
+    "contextFile": null,
+    "chunks": {
+      "ec8f8f": {
+        "user": "Karan Dutt",
+        "model": "claude-sonnet-4@20250514",
+        "timestamp": 1772697066,
+        "platform": "VisualStudioCode",
+        "product": "AIPP",
+        "productVersion": "2.2.0",
+        "source": "chat",
+        "type": "pasteCode",
+        "extraContext": {},
+        "hash": "cc105a04",
+        "messageId": "system_0ebfb960-1867-11f1-956e-13fc8e1aa9e0",
+        "codeBlockId": "system_0ebfb960-1867-11f1-956e-13fc8e1aa9e0-Codeblock-2",
+        "isActive": false
+      }
+    },
+    "lines": {}
+  }
+}
