+++ .slingshot/code-tagging/f7ae10ae-lock.json
@@ -1,0 +1,128 @@
+{
+  "backend/config/firebaseAdmin.js": {
+    "filename": "firebaseAdmin.js",
+    "path": "backend/config/firebaseAdmin.js",
+    "version": "1.2",
+    "contextFile": null,
+    "chunks": {
+      "c76680": {
+        "user": "Karan Dutt",
+        "model": "claude-sonnet-4@20250514",
+        "timestamp": 1772794824,
+        "platform": "VisualStudioCode",
+        "product": "AIPP",
+        "productVersion": "2.2.0",
+        "source": "chat",
+        "type": "pasteCode",
+        "extraContext": {},
+        "hash": "611fce7c",
+        "messageId": "system_9bc42280-194b-11f1-a9a0-6f3ba034df13",
+        "codeBlockId": "system_9bc42280-194b-11f1-a9a0-6f3ba034df13-Codeblock-1",
+        "isActive": true
+      },
+      "4dd402": {
+        "user": "Karan Dutt",
+        "model": "claude-sonnet-4@20250514",
+        "timestamp": 1772794860,
+        "platform": "VisualStudioCode",
+        "product": "AIPP",
+        "productVersion": "2.2.0",
+        "source": "chat",
+        "type": "pasteCode",
+        "extraContext": {},
+        "hash": "8ab26802",
+        "messageId": "system_9bc42280-194b-11f1-a9a0-6f3ba034df13",
+        "codeBlockId": "system_9bc42280-194b-11f1-a9a0-6f3ba034df13-Codeblock-2",
+        "isActive": true
+      }
+    },
+    "lines": {
+      "0": {
+        "chunkId": "4dd402",
+        "hash": "332fad37",
+        "encoded": "const%20admin%20=%20require(%22firebase-admin%22);",
+        "prevChunkIds": []
+      },
+      "16": {
+        "chunkId": "c76680",
+        "hash": "fdcc0195",
+        "encoded": "%7D;",
+        "prevChunkIds": []
+      },
+      "17": {
+        "chunkId": "4dd402",
+        "hash": "76477189",
+        "encoded": "%7D",
+        "prevChunkIds": []
+      },
+      "19": {
+        "chunkId": "4dd402",
+        "hash": "c91042e5",
+        "encoded": "if%20(!admin.apps.length)%20%7B",
+        "prevChunkIds": [
+          "c76680"
+        ]
+      },
+      "20": {
+        "chunkId": "4dd402",
+        "hash": "5c11a7c6",
+        "encoded": "admin.initializeApp(%7B",
+        "prevChunkIds": [
+          "c76680"
+        ]
+      },
+      "23": {
+        "chunkId": "4dd402",
+        "hash": "98201597",
+        "encoded": "%7D);",
+        "prevChunkIds": [
+          "c76680"
+        ]
+      },
+      "24": {
+        "chunkId": "c76680",
+        "hash": "71870443",
+        "encoded": "%7D",
+        "prevChunkIds": [
+          "c76680"
+        ]
+      },
+      "26": {
+        "chunkId": "4dd402",
+        "hash": "09dcc523",
+        "encoded": "const%20db%20=%20admin.firestore();",
+        "prevChunkIds": []
+      },
+      "27": {
+        "chunkId": "4dd402",
+        "hash": "2235952e",
+        "encoded": "const%20bucket%20=%20admin.storage().bucket();",
+        "prevChunkIds": []
+      },
+      "29": {
+        "chunkId": "c76680",
+        "hash": "ffb451c6",
+        "encoded": "module.exports%20=%20%7B",
+        "prevChunkIds": []
+      },
+      "30": {
+        "chunkId": "4dd402",
+        "hash": "915d8d36",
+        "encoded": "admin,",
+        "prevChunkIds": []
+      },
+      "31": {
+        "chunkId": "4dd402",
+        "hash": "4833e904",
+        "encoded": "db,",
+        "prevChunkIds": []
+      },
+      "32": {
+        "chunkId": "4dd402",
+        "hash": "92472e0b",
+        "encoded": "bucket,",
+        "prevChunkIds": []
+      }
+    }
+  }
+}
