+++ .slingshot/prompts/sample.prompt.md
@@ -1,0 +1,98 @@
+
+  SAMPLE LOCAL PROMPT
+  ───────────────────────────────
+  Welcome to your Slingshot Local Prompt setup!
+
+  ───────────────────────────────
+  HOW IT WORKS
+  ───────────────────────────────
+  You can add your own ".prompt.md" files under either of these folders available at the root of your repository
+  (In case you don't see these folders, you can either create them manually or follow the steps mentioned below):
+
+      • .slingshot/prompts/   ← recommended option
+      • .github/prompts/
+
+  All prompt files with the ".prompt.md" extension inside these folders are automatically available in the Prompt Select Drawer.
+
+  ───────────────────────────────
+  HOW TO CREATE A PROMPT
+  ───────────────────────────────
+  You can create a new prompt file by:
+    • Clicking on "Add/Edit Prompt File" from the Slash Command Drawer
+      OR
+    • Running the Slingshot command "Create or Open Prompt File"
+
+  ───────────────────────────────
+  HOW TO USE A PROMPT
+  ───────────────────────────────
+  You can use your local prompt by:
+    • Clicking on "Attach Prompt Commands" from the Slash Command Drawer
+      OR
+    • Typing the shortcut key "~" in the chat input box
+
+  The filename of your prompt will act as the command name.
+  You can use multiple local prompts at a time, but only one non-local prompt.
+
+  ───────────────────────────────
+  USING PLACEHOLDERS AND VARIABLES
+  ───────────────────────────────
+  You can reference the following VSCode workspace variables in your prompts: 
+  When creating custom prompts, users can reference the following workspace variables to dynamically insert context-specific information:
+
+- ${userHome}: Path of the user's home folder.
+- ${workspaceFolder}: Path of the folder opened in VS Code.
+- ${workspaceFolderBasename}: Name of the folder opened in VS Code without any slashes (/).
+- ${file}: Currently opened file.
+- ${fileWorkspaceFolder}: Currently opened file's workspace folder.
+- ${relativeFile}: Currently opened file relative to workspaceFolder.
+- ${relativeFileDirname}: Currently opened file's dirname relative to workspaceFolder.
+- ${fileBasename}: Currently opened file's basename.
+- ${fileBasenameNoExtension}: Currently opened file's basename with no file extension.
+- ${fileExtname}: Currently opened file's extension.
+- ${fileDirname}: Currently opened file's folder path.
+- ${fileDirnameBasename}: Currently opened file's folder name.
+- ${cwd}: Task runner's current working directory upon the startup of VS Code.
+- ${lineNumber}: Currently selected line number in the active file.
+- ${columnNumber}: Currently selected column number in the active file.
+- ${selectedText}: Currently selected text in the active file.
+- ${execPath}: Path to the running VS Code executable.
+- ${pathSeparator}: Character used by the operating system to separate components in file paths.
+
+  Additionally, you can define custom placeholders such as "{file_name}". These placeholders can be replaced with specific values when the prompt is used.
+
+  **Rules for Placeholders:**
+  - Do not include special characters like "?", "*", "+", "^", "|", "&", "!", "@", "#", "$", "%", "^", "(", ")", "=", "[", "]", ";", ":", "'", """, "<", ">", ",", ".", "/", "", "~", and "" ` "" within placeholders.
+  - Do not use double curly brackets "{{}}".
+  - Ensure placeholders are clearly defined and meaningful.
+
+  ───────────────────────────────
+  REFERENCING PROMPTS AND FILES
+  ───────────────────────────────
+  You can reference other local prompts and files within a prompt: 
+    • Reference prompts using ~{promptname} syntax.
+    • Reference files using #{path/to/file.txt} syntax.
+
+  ───────────────────────────────
+  SAMPLE PROMPT
+  ───────────────────────────────
+  For example:
+
+      My Custom Code Generation Prompt
+      --------------------------------
+      ## Follow all the below best practices while generating code
+
+      - Ensure code readability with meaningful variable and function names
+      - Maintain consistent formatting and indentation
+      - Avoid deep nesting and keep functions short and focused
+      - Handle errors and exceptions gracefully
+      - Write efficient and performant logic
+      - Add comments where necessary for complex logic
+      - Follow security best practices and avoid hardcoding secrets
+      - Include necessary input validations and edge case handling
+      - Prefer reusability and modularity over duplication
+      - Follow project-specific coding conventions and linting rules
+
+  ───────────────────────────────
+  TIP
+  ───────────────────────────────
+  Keep your prompts clear, short, and well-structured for the best results.
